Fix bugs associated with default data storage by moving to codable data storage
