+++ Todoey/Model/Item.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+struct Item: Codable {
+    var name: String = ""
+    var done: Bool = false
+}
